+++ R/spatial_vfold_cv.R
@@ -99,11 +99,21 @@ spatial_buffer_vfold_cv <- function(data,
     ...
   )
 
+  if (!missing(strata)) {
+    strata <- tidyselect::vars_select(names(data), {{ strata }})
+    if (length(strata) == 0) strata <- NULL
+  }
+
+  if (!is.null(strata)) names(strata) <- NULL
   cv_att <- list(v = v,
                  repeats = repeats,
-                 strata = !is.null(strata),
-                 radius = radius,
-                 buffer = buffer)
+                 strata  = strata,
+                 breaks  = breaks,
+                 pool    = pool,
+                 # Set radius and buffer to 0 if NULL or negative
+                 # This enables rsample::reshuffle_rset to work
+                 radius  = min(c(radius, 0)),
+                 buffer  = min(c(buffer, 0)))
 
   if ("sf" %in% class(data)) {
     rset_class <- c("spatial_buffer_vfold_cv", "spatial_rset", "rset")