+++ NEWS.md
@@ -1,5 +1,11 @@
 # spatialsample (development version)
 
+* `spatial_buffer_vfold_cv()` has had some attribute changes to match `rsample`:
+  * `strata` attribute is now the name of the column used for stratification, 
+     or not set if there was no stratification.
+  * `pool` and `breaks` have been added as attributes
+  * `radius` and `buffer` are now set to 0 if they were passed as `NULL`.
+
 # spatialsample 0.2.0
 
 ## New features